move gnosis safe verification client side
